fonction de traçage de ligne AA et modif makefile pour QEMU

+++ sysif/src/graphics.h
@@ -4,4 +4,4 @@
 #include "util.h"
 
 // Permet de tracer une ligne entre deux points
-void drawLine(uint32_t x1, uint32_t y1, uint32_t x2, uint32_t y2);
+void drawLine(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2);
